Paging back to beginning sets hasNextPage to true
When querying a connection with first, then to the next "page" with
first+after, and then back to the previous "page" with last+before
(which is the first page), the `hasNextPage` field was set to false.

The expectation is that it would be set to true (when there are more
pages). Without this change, you could page forward then back, and get
into a situation where `hasPreviousPage` and `hasNextPage` are both
false, when `hasNextPage` should be true.

Fixes https://github.com/rmosolgo/graphql-ruby/issues/1318