Reduce allocations in `apply_hooks`
`apply_hooks` is an extremely hot method.  `inject` allocates an object
to keep track of the memo used for the "fold" operation.  We don't
really need a fold operation in this case, so I converted it to an
`each`.  This kills two birds with one stone in that:

1. There are zero allocations
2. We actually short circuit iteration if `SKIP` is returned

The previous code would loop through all elements every time, even if
one of them returned `SKIP` and the subsequent procs would no longer be
called.

This patch decreases our boot time allocations by about 10 million:

```
[aaron@tc-lan-adapter ~/g/github (master)]$ env FASTDEV=1 bin/rails r 'p GC.stat(:total_allocated_objects)' ^ /dev/null
42948624
[aaron@tc-lan-adapter ~/g/github (master)]$ env FASTDEV=1 bin/rails r 'p GC.stat(:total_allocated_objects)' ^ /dev/null
32966793
[aaron@tc-lan-adapter ~/g/github (master)]$ ruby -e'p 42948624 - 32966793'
9981831
```