Allow resolve_type to return resolved object
Often abstract types in GraphQL are be backed by "wrapped" domain
objects such as unions, ADTs, etc. Currently `resolve_type` must return
a GraphQL type only even though the wrapped domain object ideally should
be unwrapped at this point as well instead of passing through to the next
resolver.

This allows abstract type's `resolve_type` to return an optional
"resolved"/unwrapped object in addition the resolved type. To optionally
return a new object value, an array can be returned as `[type, object]`.

This is backwards compatible as well and the existing validation still
applies to verify the first item in the array is a GraphQL type.