Assert that query is a string on GraphQL::Query#initialize
This commit aims to solve the issue #3619 - providing a query argument
of an incorrect type would yield a `NoMethodError` due to the incorrect
type argument to be allowed through to the lexer (which assumes it to be
a string).

As this parameter is, more often than not, feed from an external source
(e.g. a web request), this commit assumes a more defensive approach and
ensures that when the `GraphQL::Query` is initialized, the query string
is indeed of the correct type (a string).

As discussed in the issue, an `ArgumentError` is raised when a type
mismatch is identified.